# New Cache API for gathering statistics