--- a/cache/sharded_cache.cc
+++ b/cache/sharded_cache.cc
@@ -9,8 +9,6 @@
 
 #include "cache/sharded_cache.h"
 
-#include <bits/stdint-uintn.h>
-
 #include <cstdint>
 #include <memory>