--- a/cache/sharded_cache.cc
+++ b/cache/sharded_cache.cc
@@ -9,6 +9,7 @@
 
 #include "cache/sharded_cache.h"
 
+#include <algorithm>
 #include <cstdint>
 #include <memory>